+++ lib/widgets/add_transaction_form_income.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/cupertino.dart';
@@ -40,7 +38,7 @@ class _AddTransactionFormIncomeState extends State<AddTransactionFormIncome> {
       });
       final user = FirebaseAuth.instance.currentUser;
 
-      var amount = double.parse(amountEditController.text);
+      var amount = int.parse(amountEditController.text);
       DateTime date = DateTime.now();
 
       var id = uid.v4();
@@ -51,9 +49,9 @@ class _AddTransactionFormIncomeState extends State<AddTransactionFormIncome> {
           .doc(user!.uid)
           .get();
 
-      double remainingAmount = userDoc['remainingAmount'].toDouble();
-      double totalCredit = userDoc['totalCredit'].toDouble();
-      double totalDebit = userDoc['totalDebit'].toDouble();
+      int remainingAmount = userDoc['remainingAmount'];
+      int totalCredit = userDoc['totalCredit'];
+      int totalDebit = userDoc['totalDebit'];
 
       if (type == 'Credit') {
         remainingAmount += amount;
